+++ CHANGELOG.md
@@ -1,5 +1,7 @@
-## 0.2.0 / 2023-02-03
-  * Returns an array of hashes instead of a collection of objects.
+## 0.2.0 / 2023-02-04
+  * Returns Array[APage] instead of a collection of different types of objects.
+  * Converted the plugin to a `:site, :post_read` hook instead of a tag,
+    so explicit initialization is no longer required.
 
 ## 0.1.1 / 2022-04-27
   * Changed invocation from a `:site` hook to an idempotent method invocation.